# discoverer-neo/docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Discoverer Neo API
  description: >-
    REST API for Oracle Discoverer Neo metadata management backend. Manage data sources, business
    areas, folders, items, joins, hierarchies, and custom functions.
  version: 0.1.0
  contact:
    name: Discoverer Neo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token returned by POST /api/auth/login
...
paths:
...
  /api/maps/{id}/runs:
    post:
      tags:
        - Map Runs
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Default Response
  /api/runs:
    get:
      tags:
        - Map Runs
      parameters:
        - schema:
            type: string
            format: uuid
          in: query
          name: mapId
          required: false
        - schema:
            type: string
            enum:
              - QUEUED
              - RUNNING
              - COMPLETED
              - FAILED
              - CANCELLED
          in: query
          name: status
          required: false
        - schema:
            type: string
            enum:
              - LIVE
              - SCHEDULED
          in: query
          name: kind
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 200
          in: query
          name: limit
          required: false
        - schema:
            type: boolean
          in: query
          name: all
          required: false
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Default Response
  /api/runs/{id}:
    get:
      tags:
        - Map Runs
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Default Response
    delete:
      tags:
        - Map Runs
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Default Response
  /api/runs/{id}/rows:
    get:
      tags:
        - Map Runs
      parameters:
        - schema:
            type: integer
            minimum: 0
          in: query
          name: offset
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 1000
          in: query
          name: limit
          required: false
        - schema:
            type: string
            format: uuid
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Default Response
...
servers:
  - url: http://localhost:3000
    description: Local development
...